Designed kernel and enhance interfaces for the satellite component
