+++ CHANGELOG.md
@@ -8,6 +8,19 @@ the following form: YYYY.0M.0D.
 
 ## [Unreleased]
 
+## [2025.03.09]
+
+### Added
+
+- Designed kernel and enhanced interfaces for SatelliteTracker component
+
+### Updated
+
+- Changed design to include implementation for calculating orbital period
+- Updated implementation details for willCollide method
+- Introduced three final constants which is required for data manipulation in enhanced and kernel methods
+
+
 ## [2024.12.30]
 
 - Added table-based rubrics to all 6 parts of the project
