--- src/components/satellite/Satellite.java
@@ -1,5 +1,0 @@
-package components.satellitetracker;
-
-public interface SatelliteTracker extends SatelliteTrackerKernel {
-
-}
